code/requirements.txt
flask
psycopg2-binary
pandas
cdflib
numpy
scikit-learn
tensorflow
flask-cors
flask-login
python-dotenv
dotenv
requests
torch
matplotlib
